Mejorado los mensajes al usuario. El monto a convertir admite números grandes.

+++ src/Conversion.java
@@ -9,5 +9,5 @@ public record Conversion(
         String base_code,
         String target_code,
         double conversion_rate,
-        long conversion_result) {
+        double conversion_result) {
 }
